# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ShiguReader API",
    "description": "ShiguReader backend API — 类似 FastAPI 的交互式文档",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/thumbnail": {
      "get": {
        "operationId": "getThumbnail",
        "summary": "按实体名获取缩略图（tag/author/coser）",
        "tags": [
          "Thumbnail"
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "tag"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "author"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "coser"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "type",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "name",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}